[fixup] Move tests; remove redundant null check
