--- buffer/src/test/java/io/netty/buffer/CompositeByteBufTest.java
@@ -1,59 +1,0 @@
-/*
- * Copyright 2018 The Netty Project
- *
- * The Netty Project licenses this file to you under the Apache License,
- * version 2.0 (the "License"); you may not use this file except in compliance
- * with the License. You may obtain a copy of the License at:
- *
- *   http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- */
-
-package io.netty.buffer;
-
-import org.junit.Test;
-
-import static org.junit.Assert.*;
-
-public class CompositeByteBufTest {
-
-    private static final ByteBufAllocator ALLOC = UnpooledByteBufAllocator.DEFAULT;
-
-    @Test
-    public void testComponentsLessThanLowerBound() {
-        try {
-            new CompositeByteBuf(ALLOC, true, 0);
-            fail();
-        } catch (IllegalArgumentException e) {
-            assertEquals("maxNumComponents: 0 (expected: >= 1)", e.getMessage());
-        }
-    }
-
-    @Test
-    public void testComponentsEqualToLowerBound() {
-        assertCompositeBufCreated(1);
-    }
-
-    @Test
-    public void testComponentsGreaterThanLowerBound() {
-        assertCompositeBufCreated(5);
-    }
-
-    /**
-     * Assert that a new {@linkplain CompositeByteBuf} was created successfully with the desired number of max
-     * components.
-     */
-    private static void assertCompositeBufCreated(int expectedMaxComponents) {
-        CompositeByteBuf buf = new CompositeByteBuf(ALLOC, true, expectedMaxComponents);
-
-        assertNotNull(buf);
-        assertEquals(expectedMaxComponents, buf.maxNumComponents());
-
-        assertTrue(buf.release());
-    }
-}
